refactor: stress tests and benchmarks load dist, not src
Stress workers and benchmark setup exercise the shipped artifact, so keep them
on the built dist: revert stress-test/workers to import ../../dist/index.mjs
(benchmarks already do) and restore the stress config's dist version banner.
Both stress.yml and benchmark.yml build first.

Unit-test workers/fixtures continue to load src directly (no build step); the
src .ts import extensions remain for them.
